revert(preview-file): enhance desktop screenshot tool to include file URL and MIME type in response
