+++ src/qwenpaw/agents/tools/file_io.py
@@ -22,6 +22,25 @@ from ...constant import WORKING_DIR, TRUNCATION_NOTICE_MARKER
 from ...runtime.tool_registry import tool_descriptor
 
 
+def _path_to_file_url(path: str) -> str:
+    """Convert a local file path to a ``file://`` URL.
+
+    Does NOT percent-encode non-ASCII characters because agentscope's
+    DashScope formatter extracts the local path from the URL without
+    ``unquote()``, causing ``FileNotFoundError`` for files with
+    non-ASCII names (e.g. Chinese characters).
+    """
+    abs_path = os.path.abspath(path)
+    if os.name == "nt":
+        abs_path = abs_path.replace("\\", "/")
+
+    if os.name == "nt":
+        if abs_path.startswith("//"):
+            return f"file:{abs_path}"
+        return f"file:///{abs_path}"
+    return f"file://{abs_path}"
+
+
 def _resolve_file_path(file_path: str) -> str:
     """Resolve file path: use absolute path as-is,
     resolve relative path from current workspace or WORKING_DIR.
